add support for FAT filesystem and EFI partition table

+++ filesystem/src/main/java/net/lecousin/dataformat/filesystem/MBRDetector.java
@@ -22,7 +22,7 @@ public class MBRDetector implements DataFormatDetector.OnlyHeaderNeeded {
 	public DataFormat finishDetection(Data data, byte[] header, int headerLength, long dataSize) {
 		for (int i = 0; i < 4; ++i) {
 			byte b = header[0x1BE + i * 16 + 0];
-			if (b != 0 && b != 0x80)
+			if (b != 0 && b != (byte)0x80)
 				return null;
 		}
 		return MBRDataFormat.instance;
